Fix auto-verification after oauth registration
