+++ app/models/concerns/decidim/civicrm/markable_for_deletion.rb
@@ -8,6 +8,8 @@ module Decidim
       extend ActiveSupport::Concern
 
       class_methods do
+        # rubocop:disable Rails/SkipsModelValidations
+
         def self.prepare_cleanup
           update_all(marked_for_deletion: true)
         end
@@ -15,6 +17,8 @@ module Decidim
         def self.clean_up_records
           where(marked_for_deletion: true).destroy_all
         end
+
+        # rubocop:enable Rails/SkipsModelValidations
       end
     end
   end
