+++ config/locales/ca.yml
@@ -1,0 +1,5 @@
+ca:
+  decidim:
+    authorization_handlers:
+      civicrm_verification:
+        name: Autorització amb CiViCRM
