+++ decidim-civicrm.gemspec
@@ -22,6 +22,7 @@ Gem::Specification.new do |s|
 
   s.add_dependency "decidim", "~> #{Decidim::Civicrm.version}"
   s.add_dependency "decidim-core", "~> #{Decidim::Civicrm.version}"
+  s.add_dependency "decidim-verifications", "~> #{Decidim::Civicrm.version}"
 
   s.add_development_dependency "decidim-admin", "~> #{Decidim::Civicrm.version}"
   s.add_development_dependency "decidim-dev", "~> #{Decidim::Civicrm.version}"
