+++ lib/decidim/civicrm/admin_engine.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Decidim
   module Civicrm
     # This is the engine that runs on the admin interface of decidim-civicrm.
