+++ lib/decidim/civicrm/engine.rb
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 module Decidim
   module Civicrm
     # This is the engine that runs on the public interface of decidim-civicrm.
