--- lib/generators/decidim/civicrm/templates/civicrm_verification_on_oauth.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-ActiveSupport::Notifications.subscribe(/^decidim\.user\.omniauth_registration/) do |_name, data|
-  Decidim::Civicrm::VerificationJob.perform_later(data)
-end
